+++ tests/unit/model/Model.js
@@ -2139,6 +2139,191 @@ describe('Model', () => {
     });
   });
 
+  describe('traverseAsync() and $traverseAsync()', () => {
+    let Model1;
+    let Model2;
+    let model;
+
+    beforeEach(() => {
+      Model1 = modelClass('Model1');
+      Model2 = modelClass('Model2');
+
+      Model1.relationMappings = {
+        relation1: {
+          relation: Model.HasManyRelation,
+          modelClass: Model2,
+          join: {
+            from: 'Model1.id',
+            to: 'Model2.model1Id'
+          }
+        },
+        relation2: {
+          relation: Model.BelongsToOneRelation,
+          modelClass: Model1,
+          join: {
+            from: 'Model1.id',
+            to: 'Model1.model1Id'
+          }
+        }
+      };
+    });
+
+    beforeEach(() => {
+      model = Model1.fromJson({
+        id: 1,
+        model1Id: 2,
+        relation1: [{ id: 4, model1Id: 1 }, { id: 5, model1Id: 1 }],
+        relation2: {
+          id: 2,
+          model1Id: 3,
+          relation1: [{ id: 6, model1Id: 2 }, { id: 7, model1Id: 2 }],
+          relation2: {
+            id: 3,
+            model1Id: null,
+            relation1: [
+              { id: 8, model1Id: 3 },
+              { id: 9, model1Id: 3 },
+              { id: 10, model1Id: 3 },
+              { id: 11, model1Id: 3 },
+              { id: 12, model1Id: 3 },
+              { id: 13, model1Id: 3 },
+              { id: 14, model1Id: 3 },
+              { id: 15, model1Id: 3 },
+              { id: 16, model1Id: 3 },
+              { id: 17, model1Id: 3 },
+              { id: 18, model1Id: 3 },
+              { id: 19, model1Id: 3 },
+              { id: 20, model1Id: 3 },
+              { id: 21, model1Id: 3 },
+              { id: 22, model1Id: 3 },
+              { id: 23, model1Id: 3 },
+              { id: 24, model1Id: 3 },
+              { id: 25, model1Id: 3 }
+            ]
+          }
+        }
+      });
+    });
+
+    it('traverseAsync(modelArray, traverser) should traverse through the relation tree', () => {
+      let model1Ids = [];
+      let model2Ids = [];
+
+      return Model1.traverseAsync([model], model => {
+        return new Promise(resolve => {
+          setTimeout(() => {
+            if (model instanceof Model1) {
+              model1Ids.push(model.id);
+            } else if (model instanceof Model2) {
+              model2Ids.push(model.id);
+            }
+            resolve();
+          }, 5);
+        });
+      }).then(() => {
+        expect(_.sortBy(model1Ids)).to.eql([1, 2, 3]);
+        expect(_.sortBy(model2Ids)).to.eql(_.range(4, 26));
+      });
+    });
+
+    it('traverseAsync(singleModel, traverser) should traverse through the relation tree', () => {
+      let model1Ids = [];
+      let model2Ids = [];
+
+      return Model1.traverseAsync(model, model => {
+        return new Promise(resolve => {
+          setTimeout(() => {
+            if (model instanceof Model1) {
+              model1Ids.push(model.id);
+            } else if (model instanceof Model2) {
+              model2Ids.push(model.id);
+            }
+            resolve();
+          }, 5);
+        });
+      }).then(() => {
+        expect(_.sortBy(model1Ids)).to.eql([1, 2, 3]);
+        expect(_.sortBy(model2Ids)).to.eql(_.range(4, 26));
+      });
+    });
+
+    it('traverseAsync callback should be passed the model, its parent (if any) and the relation it is in (if any)', () => {
+      return Model1.traverseAsync([model], (model, parent, relationName) => {
+        if (model instanceof Model1) {
+          if (model.id === 1) {
+            expect(parent).to.equal(null);
+            expect(relationName).to.equal(null);
+          } else if (model.id === 2) {
+            expect(parent.id).to.equal(1);
+            expect(relationName).to.equal('relation2');
+          } else if (model.id === 3) {
+            expect(parent.id).to.equal(2);
+            expect(relationName).to.equal('relation2');
+          } else {
+            throw new Error('should never get here');
+          }
+        } else if (model instanceof Model2) {
+          if (model.id >= 4 && model.id <= 5) {
+            expect(parent).to.be.a(Model1);
+            expect(parent.id).to.equal(1);
+            expect(relationName).to.equal('relation1');
+          } else if (model.id >= 6 && model.id <= 7) {
+            expect(parent).to.be.a(Model1);
+            expect(parent.id).to.equal(2);
+            expect(relationName).to.equal('relation1');
+          } else if (model.id >= 8 && model.id <= 25) {
+            expect(parent).to.be.a(Model1);
+            expect(parent.id).to.equal(3);
+            expect(relationName).to.equal('relation1');
+          } else {
+            throw new Error('should never get here');
+          }
+        }
+      });
+    });
+
+    it('traverseAsync(ModelClass, model, traverser) should traverse through all ModelClass instances in the relation tree', () => {
+      let model1Ids = [];
+      let model2Ids = [];
+
+      return Model1.traverseAsync(Model2, model, model => {
+        model2Ids.push(model.id);
+      })
+        .then(() => {
+          return Model1.traverseAsync(Model1, model, model => {
+            model1Ids.push(model.id);
+          });
+        })
+        .then(() => {
+          expect(_.sortBy(model1Ids)).to.eql([1, 2, 3]);
+          expect(_.sortBy(model2Ids)).to.eql(_.range(4, 26));
+        });
+    });
+
+    it('$traverseAsync(traverser) should traverse through the relation tree', () => {
+      let model1Ids = [];
+      let model2Ids = [];
+
+      return model
+        .$traverseAsync(model => {
+          return new Promise(resolve => {
+            setTimeout(() => {
+              if (model instanceof Model1) {
+                model1Ids.push(model.id);
+              } else if (model instanceof Model2) {
+                model2Ids.push(model.id);
+              }
+              resolve();
+            }, 5);
+          });
+        })
+        .then(() => {
+          expect(_.sortBy(model1Ids)).to.eql([1, 2, 3]);
+          expect(_.sortBy(model2Ids)).to.eql(_.range(4, 26));
+        });
+    });
+  });
+
   it('$validate should run hooks and strip relations', () => {
     let Model1 = modelClass('Model1');